+++ src-bin/distrib.ml
@@ -12,7 +12,7 @@ let lint_distrib pkg ~dir =
 
 let build_distrib pkg ~dir =
   Logs.app (fun m -> m "@.Building distrib in %a" Fpath.pp dir);
-  let args = Cmd.(v "--installer" % "false" % "--vcs" % "false") in
+  let args = Cmd.(v "--pinned" % "false" % "--vcs" % "false") in
   let out = OS.Cmd.out_string in
   Topkg_care.Pkg.build pkg ~dir ~args ~out >>= function
   | (_, (_, `Exited 0)) ->