+++ src/topkg_conf.mli
@@ -62,6 +62,7 @@ val of_cli_args :
 val pkg_name : t -> string
 val build_dir : t -> Topkg_fpath.t
 val vcs : t -> bool
+val pinned : t -> bool
 val installer : t -> bool
 
 type build_context = [`Dev | `Distrib | `Pin ]