+++ CHANGES.md
@@ -1,3 +1,10 @@
+
+
+- Change pin build detection (#44). This changes the way package
+  should be build in in OPAM files. Substitute `"--installer" "true"`
+  with `"--pinned" "%{pinned}%"` in your package descriptions. The
+  `--installer` option is deprecated and has no longer any effect.
+
 v0.7.2 2016-06-02 Cambridge (UK)
 --------------------------------